+++ jax/lax/__init__.py
@@ -46,11 +46,6 @@ from jax._src.lax.lax import (
   atanh as atanh,
   atanh_p as atanh_p,
   batch_matmul as batch_matmul,
-  bessel_i0e as bessel_i0e,
-  bessel_i0e_p as bessel_i0e_p,
-  bessel_i1e as bessel_i1e,
-  bessel_i1e_p as bessel_i1e_p,
-  betainc as betainc,
   bitcast_convert_type as bitcast_convert_type,
   bitcast_convert_type_p as bitcast_convert_type_p,
   bitwise_and as bitwise_and,
@@ -87,8 +82,6 @@ from jax._src.lax.lax import (
   cosh_p as cosh_p,
   create_token as create_token,
   create_token_p as create_token_p,
-  digamma as digamma,
-  digamma_p as digamma_p,
   div as div,
   div_p as div_p,
   dot as dot,
@@ -98,12 +91,6 @@ from jax._src.lax.lax import (
   dtypes as dtypes,
   eq as eq,
   eq_p as eq_p,
-  erf as erf,
-  erf_inv as erf_inv,
-  erf_inv_p as erf_inv_p,
-  erf_p as erf_p,
-  erfc as erfc,
-  erfc_p as erfc_p,
   exp as exp,
   exp_p as exp_p,
   expand_dims as expand_dims,
@@ -117,12 +104,6 @@ from jax._src.lax.lax import (
   ge_p as ge_p,
   gt as gt,
   gt_p as gt_p,
-  igamma as igamma,
-  igamma_grad_a as igamma_grad_a,
-  igamma_grad_a_p as igamma_grad_a_p,
-  igamma_p as igamma_p,
-  igammac as igammac,
-  igammac_p as igammac_p,
   imag as imag,
   imag_p as imag_p,
   infeed as infeed,
@@ -136,8 +117,6 @@ from jax._src.lax.lax import (
   itertools as itertools,
   le as le,
   le_p as le_p,
-  lgamma as lgamma,
-  lgamma_p as lgamma_p,
   log as log,
   log1p as log1p,
   log1p_p as log1p_p,
@@ -171,8 +150,6 @@ from jax._src.lax.lax import (
   population_count_p as population_count_p,
   pow as pow,
   pow_p as pow_p,
-  random_gamma_grad as random_gamma_grad,
-  random_gamma_grad_p as random_gamma_grad_p,
   real as real,
   real_p as real_p,
   reciprocal as reciprocal,
@@ -187,7 +164,6 @@ from jax._src.lax.lax import (
   reduce_prod_p as reduce_prod_p,
   reduce_sum_p as reduce_sum_p,
   reduce_xor_p as reduce_xor_p,
-  regularized_incomplete_beta_p as regularized_incomplete_beta_p,
   rem as rem,
   rem_p as rem_p,
   reshape as reshape,
@@ -245,6 +221,32 @@ from jax._src.lax.lax import (
   unop_dtype_rule as unop_dtype_rule,
   xor_p as xor_p,
   zeros_like_array as zeros_like_array,
+)
+from jax._src.lax.special import (
+  bessel_i0e as bessel_i0e,
+  bessel_i0e_p as bessel_i0e_p,
+  bessel_i1e as bessel_i1e,
+  bessel_i1e_p as bessel_i1e_p,
+  betainc as betainc,
+  digamma as digamma,
+  digamma_p as digamma_p,
+  erf as erf,
+  erfc as erfc,
+  erfc_p as erfc_p,
+  erf_inv as erf_inv,
+  erf_inv_p as erf_inv_p,
+  erf_p as erf_p,
+  igamma as igamma,
+  igammac as igammac,
+  igammac_p as igammac_p,
+  igamma_grad_a as igamma_grad_a,
+  igamma_grad_a_p as igamma_grad_a_p,
+  igamma_p as igamma_p,
+  lgamma as lgamma,
+  lgamma_p as lgamma_p,
+  random_gamma_grad as random_gamma_grad,
+  random_gamma_grad_p as random_gamma_grad_p,
+  regularized_incomplete_beta_p as regularized_incomplete_beta_p,
 )
 from jax._src.lax.slicing import (
   GatherDimensionNumbers as GatherDimensionNumbers,
